+++ docs/conf.py
@@ -65,7 +65,8 @@ else:
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.intersphinx', 'sphinx.ext.todo']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.intersphinx', 'sphinx.ext.todo',
+              'sphinx.ext.extlinks']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = PathList(['_templates'])
@@ -289,6 +290,15 @@ intersphinx_mapping = {
     'python': ('http://docs.python.org/3/', None),
     'setuptools': ('https://pythonhosted.org/setuptools/', None)
 }
+
+
+extlinks = {
+    'issue': ('https://github.com/earthreader/libearth/issues/%s', 'issue #'),
+    'commit': ('https://github.com/earthreader/libearth/commit/%s', ''),
+    'branch': ('https://github.com/earthreader/libearth/compare/master...%s',
+               '')
+}
+
 
 
 # IronPython runtime mock