+++ docs/changes.rst
@@ -6,6 +6,9 @@ Version 0.1.1
 
 To be released.
 
+- Added a workaround for thread unsafety :func:`time.strftime()` on CPython.
+  See http://bugs.python.org/issue7980 as well.  [:issue:`32`]
+
 
 Version 0.1.0
 -------------