+++ docs/changes.rst
@@ -16,6 +16,15 @@ To be released.
   which is simply used as it is.
 
 
+Version 0.1.1
+-------------
+
+To be released.
+
+- Added a workaround for thread unsafety :func:`time.strftime()` on CPython.
+  See http://bugs.python.org/issue7980 as well.  [:issue:`32`]
+
+
 Version 0.1.0
 -------------